+++ src/features/user-meta-data/hooks/use-regions-query.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getRegions } from '../api/client-services';
-import { STALE_TIME } from '@/constants/time';
+import { STALE_TIME } from '@/constants/time-constants';
 import { META_DATA_QUERY_KEY } from '@/constants/query-key';
 const { A_DAY } = STALE_TIME;
 const { REGIONS } = META_DATA_QUERY_KEY;
